+++ app/logic.js
@@ -1,9 +1,10 @@
+function TicTacToeState(grid) {
+  this.row1 = grid.row1;
+  this.row2 = grid.row2;
+  this.row3 = grid.row3;
+}
+
 var logic = {
-  TicTacToeState(grid) {
-    this.row1 = grid.row1;
-    this.row2 = grid.row2;
-    this.row3 = grid.row3;
-  },
   /** Returns an array of coordinates of remaining blank squares
     * @param {object} grid The state of the grid
     * @return {array} coords The array of coordinates of legal moves
@@ -23,8 +24,47 @@ var logic = {
     }
     return coords;
   },
-  /** Returns a 3 * 3 array of arrays, first coordinate corresponding to row
-    * number, second coordinate listing number of blank squares, number of O
+  checkWin(grid) {
+    var i;
+    var j;
+    var product;
+    var rowName = "";
+    // Check rows
+    for (i = 0; i < 3; i++) {
+      rowName = "row" + (i + 1).toString();
+      product = 1;
+      for (j = 0; j < 3; j++) {
+        product *= grid[rowName][j];
+      }
+      if (product === 1 ||  product === 8) {
+        this.handleWin();
+        return true;
+      }
+    }
+    // Check cols
+    for (i = 0; i < 3; i++) {
+      product = 1;
+      for (j = 0; j < 3; j++) {
+        rowName = "row" + (j + 1).toString();
+        product *= grid[rowName][i];
+      }
+      if (product === 1 ||  product === 8) {
+        return true;
+      }
+    }
+    // Check diagonals
+    product = grid.row1[0] * grid.row2[1] * grid.row3[2];
+    if (product === 1 ||  product === 8) {
+      return true;
+    }
+    product = grid.row3[0] * grid.row2[1] * grid.row1[2];
+    if (product === 1 ||  product === 8) {
+      return true;
+    }
+    return false;
+  },
+  /** Returns a 3 * 3 array of arrays, for each row,
+    * lists number of blank squares, number of O
     * squares, number of X squares
     * @param {object} grid The state of the grid
     * @return {array} rowInfo Number of blank, O, and X squares for each row
